Cree ChatRepository
Modifiqué el tipo de dato Integer al id y le puse Long.
Cree el repositorio para Chat con sus respectivos métodos a usar en el Service

+++ src/main/java/com/aquienllamo/aquienllamo/model/entities/ChatEntity.java
@@ -20,7 +20,7 @@ public class ChatEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column (name = "id_chat")
-    private Integer idChat;
+    private Long idChat;
 
     @ManyToOne
     @JoinColumn(name = "id_usuario", nullable = false)
